# Tune eager fetching to speed up LoadSnapshot