`Unpause` is a bottleneck for workflow performance - p90 Unpause time is about 25s according to metrics. For users, this means that 1 in 10 workflows spend about 25s in the "Invocation is waiting for available worker" state in the UI before they even start executing.

This PR reduces this latency by tuning the eager fetching parameters. Tested on some production snapshots - this reduces the LoadSnapshot duration from ~20-27s (~1.8 Gbps peak RX throughput, measured with `iftop`) to about 10.4s (~5 Gbps peak throughput).

**Related issues**: N/A